Update guide to include info for starting service

+++ service/stop-service.sh
@@ -1,3 +1,6 @@
-PID=$(<pid.txt)
-kill -9 $PID
-rm pid.txt
+if [ -e "pid.txt" ]
+then
+	PID=$(<"pid.txt")
+	kill -9 $PID
+	rm "pid.txt"
+fi
